+++ src/edu/jhu/thrax/hadoop/features/pivot/PivotedLexicalTargetGivenSourceFeature.java
@@ -3,7 +3,7 @@ package edu.jhu.thrax.hadoop.features.pivot;
 import java.util.HashSet;
 import java.util.Set;
 
-import org.apache.hadoop.io.DoubleWritable;
+import org.apache.hadoop.io.FloatWritable;
 import org.apache.hadoop.io.Text;
 
 import edu.jhu.thrax.hadoop.datatypes.FeatureMap;
@@ -28,11 +28,11 @@ public class PivotedLexicalTargetGivenSourceFeature extends
 		return prereqs;
 	}
 
-	public DoubleWritable pivot(FeatureMap src, FeatureMap tgt) {
-		double egf = ((DoubleWritable) src.get(new Text("Lex(e|f)"))).get();
-		double fge = ((DoubleWritable) tgt.get(new Text("Lex(f|e)"))).get();
+	public FloatWritable pivot(FeatureMap src, FeatureMap tgt) {
+		float egf = ((FloatWritable) src.get(new Text("Lex(e|f)"))).get();
+		float fge = ((FloatWritable) tgt.get(new Text("Lex(f|e)"))).get();
 
-		return new DoubleWritable(egf + fge);
+		return new FloatWritable(egf + fge);
 	}
 
 	@Override
