+++ src/edu/jhu/thrax/hadoop/features/pivot/PivotedTargetPhraseGivenLHSFeature.java
@@ -3,7 +3,7 @@ package edu.jhu.thrax.hadoop.features.pivot;
 import java.util.HashSet;
 import java.util.Set;
 
-import org.apache.hadoop.io.DoubleWritable;
+import org.apache.hadoop.io.FloatWritable;
 import org.apache.hadoop.io.Text;
 
 import edu.jhu.thrax.hadoop.datatypes.FeatureMap;
@@ -26,8 +26,8 @@ public class PivotedTargetPhraseGivenLHSFeature extends NonAggregatingPivotedFea
     return prereqs;
   }
 
-  public DoubleWritable pivot(FeatureMap src, FeatureMap tgt) {
-    return new DoubleWritable(((DoubleWritable) tgt.get(new Text("p(e|LHS)"))).get());
+  public FloatWritable pivot(FeatureMap src, FeatureMap tgt) {
+    return new FloatWritable(((FloatWritable) tgt.get(new Text("p(e|LHS)"))).get());
   }
 
   @Override
